Move ExtractBlockSigningKeys into staking namespace as free functions
