+++ src/Makefile.test.include
@@ -37,7 +37,6 @@ UNITE_TESTS = \
   test/bech32_tests.cpp \
   test/better_enum_set_tests.cpp \
   test/bip32_tests.cpp \
-  test/blockchain/blockchain_behavior_tests.cpp \
   test/blockchain/blockchain_custom_parameters_tests.cpp \
   test/blockchain/blockchain_parameters_tests.cpp \
   test/blockchain_tests.cpp \
@@ -116,6 +115,7 @@ UNITE_TESTS = \
   test/snapshot/snapshot_index_tests.cpp \
   test/snapshot/state_tests.cpp \
   test/snapshot/validation_tests.cpp \
+  test/staking/proof_of_stake_tests.cpp \
   test/streams_tests.cpp \
   test/test_unite.h \
   test/test_unite.cpp \
